Make synchronous/asynchronous modules crate private and re-export necessary types
